+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- New: `--features`, `--no-default-features` and `--all-features` options are passed through to Cargo.
+
 ## 24.2.0
 
 - New: Colored output can be enabled in CI or other noninteractive situations by passing `--colors=always`, or setting `CARGO_TERM_COLOR=always`, or `CLICOLOR_FORCE=1`. Colors can similarly be forced off with `--colors=never`, `CARGO_TERM_COLOR=never`, or `NO_COLOR=1`.